Resolução do desafio.

+++ src/main/java/com/trybe/acc/java/planejamentodeviagem/Principal.java
@@ -33,7 +33,7 @@ public class Principal {
             .print("Entre com a data e o horário da partida (formato: dd/mm/aaaa hh:mm:ss): ");
         String partida = scanner.nextLine();
         System.out.print("Entre com a distância em km entre a cidade de origem e a de destino: ");
-        Double distanciaKm = scanner.nextDouble();
+        int distanciaKm = scanner.nextInt();
 
         Viagem viagem = new Viagem(partida, origem, destino, distanciaKm);
